button changes to bin symbol and colors sass section added

+++ src/components/expenses-list-item/ExpensesListItem.tsx
@@ -1,5 +1,6 @@
 import { Link, useParams } from 'react-router-dom'
 import './expense-list-item.styles.scss'
+import { FaTrashAlt } from 'react-icons/fa'
 import { formatDate } from '../../utils/format/format.utils'
 import { useAppDispatch, useAppSelector } from '../../hooks'
 import { removeExpenseFromExpenses } from '../../store/expenses/expenses.reducer'
@@ -7,7 +8,6 @@ import { removeExpenseFromCollection } from '../../utils/firebase/firebase.utils
 import { Expense } from '../../store/tricounts/tricounts.types'
 import { selectCurrentUser } from '../../store/user/user.selector'
 import { TriCountRouteParams } from '../header-list/HeaderList'
-import Button, { BUTTON_TYPE_CLASSES } from '../button/Button'
 
 interface ExpensesListItemProps {
   expense: Expense
@@ -23,14 +23,16 @@ function ExpensesListItem({ expense, symbol }: ExpensesListItemProps) {
   const dispatch = useAppDispatch()
   const handleOnClick = async (e: React.MouseEvent<HTMLButtonElement>) => {
     e.preventDefault()
-    dispatch(removeExpenseFromExpenses(id))
-    if (user) {
-      await removeExpenseFromCollection(
-        'tricounts',
-        expense,
-        user.uid,
-        tricountId
-      )
+    if (window.confirm('Are you sure you want to delete ?')) {
+      dispatch(removeExpenseFromExpenses(id))
+      if (user) {
+        await removeExpenseFromCollection(
+          'tricounts',
+          expense,
+          user.uid,
+          tricountId
+        )
+      }
     }
   }
 
@@ -45,13 +47,8 @@ function ExpensesListItem({ expense, symbol }: ExpensesListItemProps) {
           </p>
         </div>
         <div>
-          <Button
-            buttonType={BUTTON_TYPE_CLASSES.base}
-            type="button"
-            onClick={handleOnClick}
-          >
-            x
-          </Button>
+          <FaTrashAlt size={20} onClick={handleOnClick} />
+
           <div className="expenses-date">{formattedDate}</div>
           <div
             className="expenses-price"
